requirements.txt
asgiref==3.2.3
Django==3.0
djangorestframework==3.10.3
et-xmlfile==1.0.1
jdcal==1.4.1
openpyxl==3.0.2
pytz==2019.3
sqlparse==0.3.0
xlwt==1.3.0
